requirements.txt
asyncio
datetime
aiosqlite
easy-pil
python-dotenv
nextcord
pymongo
dnspython
Pillow
requests
aiohttp
datetime
traceback
sqlite3
validators
pypresence
git+https://github.com/Potatopy/Discord-iOS
git+https://github.com/Potatopy/wavelink-nextcord
